requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
cached-property==1.5.2
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
crypto==1.4.1
defusedxml==0.7.1
Django==3.2.3
django-cors-headers==3.8.0
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
isodate==0.6.0
kombu==5.0.2
lxml==4.6.3
Naked==0.1.31
persiantools==2.1.1
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycryptodome==3.10.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
selenium==3.141.0
shellescape==3.8.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
suds-py3==1.4.4.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
zeep==4.0.0
